# openapi.yaml
openapi: 3.1.0
info:
  contact:
    url: https://www.asam.net/project-detail/sovd-service-oriented-vehicle-diagnostics/
  description: Definition of ASAM SOVD API
  license:
    name: Licensing Terms for Grant of Rights to Use ASAM Products
    url: https://www.asam.net/license/
  title: Service-Oriented Vehicle Diagnostic API
  version: 1.0.0
servers:
- url: https://sovd.server/v1
...
paths:
...
  /{entity-collection}/{entity-id}/modes:
    get:
      description: This method returns all available modes defined for the given entity.
      parameters:
      - description: Path to an entity collection
        explode: false
        in: path
        name: entity-collection
        required: true
        schema:
          enum:
          - Area
          - Component
          - App
          - Function
          type: string
        style: simple
      - description: ID of an entity
        explode: false
        in: path
        name: entity-id
        required: true
        schema:
          type: string
        style: simple
      - description: "Specifies whether the response should include schema information\
          \ or not. Default: false"
        explode: true
        in: query
        name: include-schema
        required: false
        schema:
          type: boolean
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/__entity_collection___entity_id__modes_get_200_response'
          description: The request was successful
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/__any_path__docs_get_default_response'
          description: An unexpected error occurred.
      tags:
      - target-modes
...
  /{entity-collection}/{entity-id}/logs/entries:
    get:
      description: This method provides the log information aggregated by that entity.
      parameters:
      - description: Path to an entity collection
        explode: false
        in: path
        name: entity-collection
        required: true
        schema:
          enum:
          - Area
          - Component
          - App
          - Function
          type: string
        style: simple
      - description: ID of an entity
        explode: false
        in: path
        name: entity-id
        required: true
        schema:
          type: string
        style: simple
      - explode: true
        in: query
        name: severity
        required: false
        schema:
          description: The severity specifies the log-level for the given context.
          enum:
          - fatal
          - error
          - warn
          - info
          - debug
          title: severity
          type: string
        style: form
      - description: "Specifies whether the response should include schema information\
          \ or not. Default: false"
        explode: true
        in: query
        name: include-schema
        required: false
        schema:
          type: boolean
        style: form
      responses:
        "200":
          content:
            application/json:
              example:
                items:
                - timestamp: 2021-07-20T00:00:04.387819Z
                  context:
                    type: RFC5424
                    host: Linux
                    process: systemd
                    pid: 1
                  severity: info
                  msg: Closed D-Bus User Message Bus Socket
                  href: https://sovd.server/v1/components/DrivingComputer/bulk-data/logs/server.log
              schema:
                $ref: '#/components/schemas/__entity_collection___entity_id__logs_entries_get_200_response'
          description: The request was successful.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/__any_path__docs_get_default_response'
          description: An unexpected error occurred.
      tags:
      - logging
  /{entity-collection}/{entity-id}/logs/config:
    delete:
      description: This method resets the current configuration to its default.
      parameters:
      - description: Path to an entity collection
        explode: false
        in: path
        name: entity-collection
        required: true
        schema:
          enum:
          - Area
          - Component
          - App
          - Function
          type: string
        style: simple
      - description: ID of an entity
        explode: false
        in: path
        name: entity-id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: Configuration has been reset successfully.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/__any_path__docs_get_default_response'
          description: The SOVD client tried to reset a configuration which does not
            exist.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/__any_path__docs_get_default_response'
          description: An unexpected error occurred.
      tags:
      - logging
    get:
      description: "This method returns an array of configured SOVD LogConfiguration\
        \ elements, specifying the different contexts for which a specific log configuration\
        \ was created with the method Configure SOVD Logging."
      parameters:
      - description: Path to an entity collection
        explode: false
        in: path
        name: entity-collection
        required: true
        schema:
          enum:
          - Area
          - Component
          - App
          - Function
          type: string
        style: simple
      - description: ID of an entity
        explode: false
        in: path
        name: entity-id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              example:
                contexts:
                - context:
                    type: RFC5424
                    host: Linux
                    process: systemd
                    pid: 1
                  severity: info
                - context:
                    type: AUTOSAR_DLT
                    session: SES01
                    session_id: 0x0001
                    application_id: WindowControl
                    context_id: STD0
                    message_id: msg000012
                  severity: info
              schema:
                $ref: '#/components/schemas/__entity_collection___entity_id__logs_config_get_200_response'
          description: The request was successful.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/__any_path__docs_get_default_response'
          description: An unexpected error occurred.
      tags:
      - logging
    put:
      description: This method configures the log aggregation.
      parameters:
      - description: Path to an entity collection
        explode: false
        in: path
        name: entity-collection
        required: true
        schema:
          enum:
          - Area
          - Component
          - App
          - Function
          type: string
        style: simple
      - description: ID of an entity
        explode: false
        in: path
        name: entity-id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            example:
              items:
              - context:
                  type: RFC5424
                  host: Linux
                  process: systemd
                severity: warn
            schema:
              $ref: '#/components/schemas/__entity_collection___entity_id__logs_config_put_request'
        description: Request Body - Configure SOVD logging
        required: true
      responses:
        "204":
          description: The configuration was accepted.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/__any_path__docs_get_default_response'
          description: Invalid configuration provided
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/__any_path__docs_get_default_response'
          description: An unexpected error occurred.
      tags:
      - logging
...
components:
  schemas:
...
    __entity_collection___entity_id__logs_entries_get_200_response_items_inner_context_oneOf_1:
      example:
        type: AUTOSAR_DLT
        session: SES01
        session_id: 0x0001
        application_id: WindowControl
        context_id: STD0
        message_id: msg000012
      properties:
        type:
          type: string
        session:
          title: id
          type: string
        session_id:
          title: id
          type: string
        application_id:
          title: id
          type: string
        context_id:
          title: id
          type: string
        message_id:
          title: id
          type: string
      required:
      - type
      title: __entity_collection___entity_id__logs_entries_get_200_response_items_inner_context_oneOf_1
...
    __entity_collection___entity_id__logs_entries_get_200_response_items_inner:
      example:
        timestamp: 2021-07-20T00:00:04.387819Z
        context:
          type: RFC5424
          host: Linux
          process: systemd
          pid: 1
        severity: info
        msg: Closed D-Bus User Message Bus Socket
        href: https://sovd.server/v1/components/DrivingComputer/bulk-data/logs/server.log
      properties:
        timestamp:
          description: "Format: RFC3339 (UTC)"
          format: date-time
          type: string
        context:
          $ref: '#/components/schemas/__entity_collection___entity_id__logs_entries_get_200_response_items_inner_context'
        severity:
          description: The severity specifies the log-level for the given context.
          enum:
          - fatal
          - error
          - warn
          - info
          - debug
          title: severity
          type: string
        msg:
          description: The actual log message (text)
          type: string
        href:
          description: Reference to a URI for retrieving the underlying original log
            file in an application specific format
          format: uri-reference
          type: string
      required:
      - context
      - msg
      - severity
      - timestamp
      title: __entity_collection___entity_id__logs_entries_get_200_response_items_inner
...
    __entity_collection___entity_id__modes_get_200_response:
      example:
        items:
        - id: session
          name: Diagnostic session
      properties:
        items:
          description: The list of available modes of an entity.
          items:
            $ref: '#/components/schemas/__entity_collection___entity_id__modes_get_200_response_items_inner'
          type: array
        schema:
          description: "The schema definition of the response. Condition: Only provided\
            \ if the query parameter include-schema is true."
          title: schema
          type: object
      required:
      - items
      title: __entity_collection___entity_id__modes_get_200_response
...
    __entity_collection___entity_id__logs_config_get_200_response:
      example:
        contexts:
        - severity: fatal
          context:
            type: RFC5424
            host: Linux
            process: systemd
            pid: 1
        - severity: fatal
          context:
            type: RFC5424
            host: Linux
            process: systemd
            pid: 1
      properties:
        contexts:
          description: An array of LogConfiguration elements
          items:
            $ref: '#/components/schemas/__entity_collection___entity_id__logs_config_get_200_response_contexts_inner'
          title: contexts
          type: array
      required:
      - contexts
      title: __entity_collection___entity_id__logs_config_get_200_response
...
    __entity_collection___entity_id__logs_config_put_request:
      properties:
        items:
          items:
            $ref: '#/components/schemas/__entity_collection___entity_id__logs_config_get_200_response_contexts_inner'
          title: items
          type: array
      required:
      - items
      title: __entity_collection___entity_id__logs_config_put_request
...
    __entity_collection___entity_id__modes_get_200_response_items_inner:
      properties:
        id:
          description: The resource identifier of a mode of an entity.
          title: id
          type: string
        name:
          description: The name of a mode of an entity.
          title: name
          type: string
        translation_id:
          description: Translation identifier for the name
          title: translation_id
          type: string
      required:
      - id
      title: __entity_collection___entity_id__modes_get_200_response_items_inner
...
    __entity_collection___entity_id__logs_config_get_200_response_contexts_inner:
      example:
        severity: fatal
        context:
          type: RFC5424
          host: Linux
          process: systemd
          pid: 1
      properties:
        context:
          $ref: '#/components/schemas/__entity_collection___entity_id__logs_config_get_200_response_contexts_inner_context'
        severity:
          description: The severity specifies the log-level for the given context.
          enum:
          - fatal
          - error
          - warn
          - info
          - debug
          title: severity
          type: string
      required:
      - context
      - severity
      title: __entity_collection___entity_id__logs_config_get_200_response_contexts_inner
...
    __entity_collection___entity_id__logs_config_get_200_response_contexts_inner_context:
      description: Definition of the context of the log message. Parts not specified
        in the context are regarded as not set.
      oneOf:
      - $ref: '#/components/schemas/__entity_collection___entity_id__logs_entries_get_200_response_items_inner_context_oneOf'
      - $ref: '#/components/schemas/__entity_collection___entity_id__logs_entries_get_200_response_items_inner_context_oneOf_1'
      title: __entity_collection___entity_id__logs_config_get_200_response_contexts_inner_context
...
    __entity_collection___entity_id__logs_entries_get_200_response:
      example:
        schema: "{}"
        items:
        - timestamp: 2021-07-20T00:00:04.387819Z
          context:
            type: RFC5424
            host: Linux
            process: systemd
            pid: 1
          severity: info
          msg: Closed D-Bus User Message Bus Socket
          href: https://sovd.server/v1/components/DrivingComputer/bulk-data/logs/server.log
        - timestamp: 2021-07-20T00:00:04.387819Z
          context:
            type: RFC5424
            host: Linux
            process: systemd
            pid: 1
          severity: info
          msg: Closed D-Bus User Message Bus Socket
          href: https://sovd.server/v1/components/DrivingComputer/bulk-data/logs/server.log
      properties:
        items:
          items:
            $ref: '#/components/schemas/__entity_collection___entity_id__logs_entries_get_200_response_items_inner'
          title: items
          type: array
        schema:
          description: "The schema definition of the response. Condition: Only provided\
            \ if the query parameter include-schema is true."
          title: schema
          type: object
      required:
      - items
      title: __entity_collection___entity_id__logs_entries_get_200_response
...
    __entity_collection___entity_id__logs_entries_get_200_response_items_inner_context:
      description: Definition of the context of the log message identifying the source
        of the entry.
      oneOf:
      - $ref: '#/components/schemas/__entity_collection___entity_id__logs_entries_get_200_response_items_inner_context_oneOf'
      - $ref: '#/components/schemas/__entity_collection___entity_id__logs_entries_get_200_response_items_inner_context_oneOf_1'
      title: __entity_collection___entity_id__logs_entries_get_200_response_items_inner_context
...
    __any_path__docs_get_default_response:
      description: It is recommended that the response body of an error response includes
        a more detailed description of the error which helps in diagnosing the issue.
      example:
        vendor_code: vendor_code
        translation_id: translation_id
        error_code: error_code
        message: message
        parameters: "{}"
      properties:
        error_code:
          description: SOVD standardized error code
          title: error_code
          type: string
        vendor_code:
          description: "Vendor specific error code Condition: If the error_code is\
            \ vendor-specific this attribute has to be filled."
          title: vendor_code
          type: string
        message:
          description: Message describing the problem in more detail.
          title: message
          type: string
        translation_id:
          description: Identifier for translating the message
          title: translation_id
          type: string
        parameters:
          description: Key-value-list which provides additional information on the
            error
          title: parameters
          type: object
      required:
      - error_code
      - message
      title: __any_path__docs_get_default_response
...
    __entity_collection___entity_id__logs_entries_get_200_response_items_inner_context_oneOf:
      example:
        type: RFC5424
        host: Linux
        process: systemd
        pid: 1
      properties:
        type:
          type: string
        host:
          title: id
          type: string
        process:
          title: id
          type: string
        pid:
          type: integer
      required:
      - type
      title: __entity_collection___entity_id__logs_entries_get_200_response_items_inner_context_oneOf
...